Update model zoo and app ssd scripts and readme (#90)
* update

* update pom and scripts

* update readme

+++ models/build.sh
@@ -1,10 +1,20 @@
 mvn clean package
-BigDL_PYTHON=/home/jxy/code/Bigdl/spark-dl/spark/dist/target/bigdl-0.4.0-SNAPSHOT-spark-2.0.0-scala-2.11.8-all-dist/lib/bigdl-0.4.0-SNAPSHOT-python-api.zip
+# Donwload BigDL python api
+BigDL=dist-spark-2.1.1-scala-2.11.8-all-0.4.0-dist
+if [ ! -d $BigDL ]; then
+    wget https://s3-ap-southeast-1.amazonaws.com/bigdl-download/$BigDL.zip
+    unzip $BigDL.zip -d $BigDL
+fi
+# Merge BigDL python API and models python API
+BigDL_PYTHON=${BigDL}/lib/bigdl-0.4.0-python-api.zip
 cp ${BigDL_PYTHON} .
-rm -r python
+if [-d python]; then
+    rm -r python
+fi
 unzip ${BigDL_PYTHON} -d python
 unzip -n target/models-0.1-SNAPSHOT-python-api.zip -d python
 cd python
 zip -r ../target/bigdl-models-0.1-SNAPSHOT-python-api.zip *
+rm -r ../python
